feat(#302): Add Defn and Undefn terms to Factbase system
