+++ lib/factbase/terms/defn.rb
@@ -3,15 +3,26 @@
 # SPDX-FileCopyrightText: Copyright (c) 2024-2025 Yegor Bugayenko
 # SPDX-License-Identifier: MIT
 
-require_relative '../../factbase'
+require_relative 'base'
 
-# Defn terms.
-#
+# Factbase::Defn is responsible for defining new terms in the Factbase system.
 # Author:: Yegor Bugayenko (yegor256@gmail.com)
 # Copyright:: Copyright (c) 2024-2025 Yegor Bugayenko
 # License:: MIT
-module Factbase::Defn
-  def defn(_fact, _maps, _fb)
+class Factbase::Defn < Factbase::TermBase
+  # Constructor.
+  # @param [Array] operands Operands
+  def initialize(operands)
+    super()
+    @operands = operands
+  end
+
+  # Evaluate term on a fact.
+  # @param [Factbase::Fact] fact The fact
+  # @param [Array<Factbase::Fact>] maps All maps available
+  # @param [Factbase] fb Factbase to use for sub-queries
+  # @return [Object] Term definition result
+  def evaluate(_fact, _maps, _fb)
     assert_args(2)
     fn = @operands[0]
     raise "A symbol expected as first argument of 'defn'" unless fn.is_a?(Symbol)
@@ -22,16 +33,6 @@ module Factbase::Defn
     # rubocop:disable Security/Eval
     eval(e)
     # rubocop:enable Security/Eval
-    true
-  end
-
-  def undef(_fact, _maps, _fb)
-    assert_args(1)
-    fn = @operands[0]
-    raise "A symbol expected as first argument of 'undef'" unless fn.is_a?(Symbol)
-    if Factbase::Term.private_instance_methods(false).include?(fn)
-      Factbase::Term.class_eval("undef :#{fn}", __FILE__, __LINE__ - 1) # undef :foo
-    end
     true
   end
 end
